Do not erase cache every login forever, when NewCacheLocation is set a non-existing directory!
I can't believe that I have not have gotten a cache for two years
because of this... I knew I didn't, but didn't really want to fix
it while I working on the network code (AICurl) ;). It kinda made
for a good testing. Bet my ISP is glad I fixed it though.