Allow users to join a new grid with Singularity without first having to install another viewer.
I've seen this in other grids (InWorldz and OpenGrid) and is a problem
on DreamNation with every new user: they can't put on a shape or skin
until they first logged in with another viewer. See comments in code.

Tested, and now could successfully add shape, skin, eyes and hair.