Add Render muted avatars check to System->Security & Privacy Preferences
Redefaulted LiruLegacyDisplayMuteds to false due to (at least) two alpha testers noticing it and wanting it back how it is in the current release (since 1.7.0).
True was to get back to the old ways of v1, but it's okay to be false by default, better now since it's exposed.

Also irrelevant tiny voice cleanup from Spatters.